Rewrite Zoom-to-S3 code
This code was arbitrarily failing for closed caption files.  It seemed
that the Lambda was only receiving part of each file from Zoom, and the
multi-part upload was failing (EntityTooSmall). In contrast, the method
suggested at [Python program to stream data from a URL and write it to
S3 ? – All About
Tech](https://amalgjose.com/2020/08/13/python-program-to-stream-data-from-a-url-and-write-it-to-s3/)
uses `s3.upload_fileobj()`.